Detect
This change adds support for the detect part of the buildpack lifecycle.
Detection passes if either build.gradle or pom.xml exist in $APPLICATION_ROOT.
If a build system is found, it contributes the type of build system, a jvm-
application key, and an openjdk-jdk key in order to build the application.