Fail If More Than One Candidate JAR
Previously, if the buildpack detected more than one built JAR, it would
blindly select the first one enumerated by the filesystem and use that.  This
can lead to negative consequences.  This change updates the buildpack to fail
in the case of both zero candidate JARs and more than one candidate JAR.

[resolves #2]