Build to Layer
Previously when an application was built, the application directory was
removed and the created artifact was expanded to the app directory directly.
This behavior was inefficient and precluded some interesting use cases so it
was changed.  Now, after a build has been completed, the artifact is expanded
to a layer.  The original application directory is then removed and that layer
is symlinked into it's place.

Signed-off-by: Ben Hale <bhale@pivotal.io>