Do Not Symlink Application
Previously the buildpack would expand the built artifact to a layer and then
symlink the application directory to the layer.  This behavior (symlink the
root application directory) was unsupported by the lifecycle and would not
work.  This change reverts that so that the built artifact is once again
expanded to the application directory directly.

Signed-off-by: Ben Hale <bhale@pivotal.io>