Remove source not directory
Previously, after the build was complete, the buildpack would remove the
application root directory and then recreate it with the compiled contents. In
situations where that directory was un-removable (e.g. if it was a filesystem
mount like in Knative) the build would fail.  This change updates the code to
remove the contents of the directory without removing the directory itself.

[resolves #5]

Signed-off-by: Ben Hale <bhale@pivotal.io>