Use Application From Submodule
Previously, when an application was built, only a "root" application could be
detected and expanded into the application folder.  This prevented the
possibility of building a monorepo and then using one of the applications
created from it.

This change adds support for the $BP_BUILD_SYSTEM_SUBMODULE environment
variable which allows users to specify where in a monorepo, the built
application should be searched found.

Signed-off-by: Ben Hale <bhale@pivotal.io>