Customize Built Artifact
Previously, a very submodule-specific solution to selecting the compiled
application was created.  However, it turned out that there are cases where
even when a submodule was specified, multiple JARs might exist in a build
directory (source, javadocs, etc.).  Therefore this solution wasn't generally
applicable.

This change modifies the environment variable to be $BP_BUILT_ARTIFACT which
can point to exactly the built artifact (optionally using globs as long as
they select a single artifact) which solves both the submodule and multiple
JAR scenarios.

Signed-off-by: Ben Hale <bhale@pivotal.io>