Configure module artifact
Previously, a built artifact could only exist in the root target/ or
build/libs/ directories after the build completed. If it existed in any other
locations (say in a multi-module build), a user would have to specify
$BP_BUILT_ARTIFACT with the module path as well as the target/ or build/libs
directory. Given that the fully qualified path was build-system specific, this
meant that a generic front end for configuring this value was problematic.

This change allows a user to set $BP_BUILT_MODULE with a module name that is
prepended to the default built artifact glob.

[resolves #10]

Signed-off-by: Ben Hale <bhale@pivotal.io>