Only Executable JARs and WARs are candidates
Currently, when searching for a candidate JAR to use after a build, there's a
limitation that there can only be a single JAR.  This change improves the
algorithm to look only for JARs that have a Main-Class and WARs with a
WEB-INF/ directory. If there is only a single one of these candidates, that
will be the application artifact.

[resolves #9]

Signed-off-by: Ben Hale <bhale@pivotal.io>