Select Single Candidate
Previously if a build resulted in a single candidate but was not interesting
(executable JAR or standard WAR) then it would not be selected.  In the case
of JARs containing functions that were not executable, this would result in a
successful build that would not proceed.

This change updates the behavior to always select a single JAR or WAR
regardless of whether it was interesting and let another buildpack further
down the line decide to fail given what was available.

Signed-off-by: Ben Hale <bhale@pivotal.io>